+++ inbm/cloudadapter-agent/fpm-template/etc/apparmor.d/usr.bin.inbm-cloudadapter
@@ -31,6 +31,7 @@
   /etc/intel-manageability/secret/cloudadapter-agent/** r,
   /etc/intel-manageability/public/ucc-ca/ r,
   /usr/share/cloudadapter-agent/config_schema.json r,
+  /usr/share/cloudadapter-agent/node_update_schema.json r,
   /etc/intel-manageability/public/cloudadapter-agent/device_id rw,
   /etc/intel-manageability/public/mqtt-ca/mqtt-ca.crt r,
   /etc/intel-manageability/public/ucc-ca/ucc.ca.pem.crt r,
