+++ inbm/cloudadapter-agent/fpm-template/usr/share/cloudadapter-agent/node_update_schema.json
